requirements.txt
# absl-py==0.9.0
# altgraph==0.17
# astor==0.8.1
# backcall==0.1.0
beautifulsoup4==4.9.3
boto3==1.17.110
# blis==0.4.1
# cachetools==4.1.1
# catalogue==1.0.0
certifi==2019.11.28
# chardet==3.0.4
click==7.1.1
# colorama==0.4.3
# cupy-cuda102==7.8.0
# cycler==0.10.0
cymem==2.0.3
# cython==0.25
dateutils==0.6.8
# decorator==4.4.2
# en-core-web-sm==2.3.1
# fastrlock==0.5
Flask==1.1.2
Flask-JWT-Extended==3.25.0
Flask-MySQLdb==0.2.0
flask-swagger==0.2.14
flask-swagger-ui==3.36.0
# future==0.18.2
# gast==0.2.2
ghostscript==0.6
# google-pasta==0.2.0
# grpcio==1.28.1
# h5py==2.10.0
# idna==2.9
# imageio==2.9.0
# importlib-metadata==1.7.0
imutils==0.5.3
ipython==7.13.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
# jedi==0.16.0
Jinja2==2.11.1
# joblib==0.16.0
# Keras==2.4.3
# Keras-Applications==1.0.8
# Keras-Preprocessing==1.1.0
# kiwisolver==1.2.0
# macholib==1.14
# Markdown==3.2.1
# MarkupSafe==1.1.1
# matplotlib==3.2.2
mock==4.0.2
murmurhash==1.0.2
mysql-connector==2.2.9
mysql-connector-python==8.0.23
# mysqlclient==1.4.3
# networkx==2.4
# nlpaug==1.1.0
nltk==3.4.5
numpy>=1.19.3
oauthlib==3.1.0
opencv-python==4.5.2.54
opt-einsum==3.3.0
pandas==1.1.5
# parso==0.6.2
pdftotext==2.1.4
# pefile==2019.4.18
pickleshare==0.7.5
Pillow==7.2.0
# plac==1.1.3
preshed==3.0.2
# prompt-toolkit==3.0.5
# protobuf==3.11.3
# pyasn1==0.4.8
# pyasn1-modules==0.2.8
Pygments==2.6.1
# PyJWT==1.7.1
# pyparsing==2.4.6
pytesseract==0.3.7
python-dateutil==2.8.1
pytz==2019.3
# PyWavelets==1.1.1
# pywin32-ctypes==0.2.0
PyYAML==5.3.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.6
regex
schedule==1.0.0
scikit-image==0.17.2
scipy==1.5.4
scikit-learn==0.22.2.post1

# seqeval==0.0.12
six==1.14.0
# soupsieve==2.0.1
spacy==2.3.2
# srsly==1.0.2
# tabulate==0.8.7
# termcolor==1.1.0
thinc==7.4.1
threadpoolctl==2.1.0
tifffile==2020.7.4
tqdm==4.45.0
# traitlets==4.3.3
urllib3==1.25.8
# wasabi==0.8.0
# wcwidth==0.1.9
Werkzeug==1.0.1
# wrapt==1.12.1
# zipp==3.1s.0
